+++ declarations/ValidationError.d.ts
@@ -2,9 +2,9 @@ export default ValidationError;
 export type JSONSchema6 = import('json-schema').JSONSchema6;
 export type JSONSchema7 = import('json-schema').JSONSchema7;
 export type Schema =
-  | import('json-schema').JSONSchema4
-  | import('json-schema').JSONSchema6
-  | import('json-schema').JSONSchema7;
+  | (import('json-schema').JSONSchema4 & import('./validate').Extend)
+  | (import('json-schema').JSONSchema6 & import('./validate').Extend)
+  | (import('json-schema').JSONSchema7 & import('./validate').Extend);
 export type ValidationErrorConfiguration = {
   name?: string | undefined;
   baseDataPath?: string | undefined;
@@ -31,9 +31,9 @@ declare class ValidationError extends Error {
       children?: import('ajv').ErrorObject[] | undefined;
     })[],
     schema:
-      | import('json-schema').JSONSchema4
-      | import('json-schema').JSONSchema6
-      | import('json-schema').JSONSchema7,
+      | (import('json-schema').JSONSchema4 & import('./validate').Extend)
+      | (import('json-schema').JSONSchema6 & import('./validate').Extend)
+      | (import('json-schema').JSONSchema7 & import('./validate').Extend),
     configuration?: import('./validate').ValidationErrorConfiguration
   );
   /** @type {Array<SchemaUtilErrorObject>} */
@@ -53,9 +53,9 @@ declare class ValidationError extends Error {
   getSchemaPart(
     path: string
   ):
-    | import('json-schema').JSONSchema4
-    | import('json-schema').JSONSchema6
-    | import('json-schema').JSONSchema7;
+    | (import('json-schema').JSONSchema4 & import('./validate').Extend)
+    | (import('json-schema').JSONSchema6 & import('./validate').Extend)
+    | (import('json-schema').JSONSchema7 & import('./validate').Extend);
   /**
    * @param {Schema} schema
    * @param {Array<Object>} prevSchemas
@@ -63,9 +63,10 @@ declare class ValidationError extends Error {
    */
   formatSchema(
     schema:
-      | import('json-schema').JSONSchema4
-      | import('json-schema').JSONSchema6
-      | import('json-schema').JSONSchema7,
+      | (import('json-schema').JSONSchema4 & import('./validate').Extend)
+      | (import('json-schema').JSONSchema6 & import('./validate').Extend)
+      | (import('json-schema').JSONSchema7 & import('./validate').Extend),
+    logic?: boolean,
     prevSchemas?: Object[]
   ): string;
   /**
@@ -76,12 +77,13 @@ declare class ValidationError extends Error {
    */
   getSchemaPartText(
     schemaPart?:
-      | import('json-schema').JSONSchema4
-      | import('json-schema').JSONSchema6
-      | import('json-schema').JSONSchema7
+      | (import('json-schema').JSONSchema4 & import('./validate').Extend)
+      | (import('json-schema').JSONSchema6 & import('./validate').Extend)
+      | (import('json-schema').JSONSchema7 & import('./validate').Extend)
       | undefined,
     additionalPath?: boolean | string[] | undefined,
-    needDot?: boolean | undefined
+    needDot?: boolean | undefined,
+    logic?: boolean
   ): string;
   /**
    * @param {Schema=} schemaPart
@@ -89,9 +91,9 @@ declare class ValidationError extends Error {
    */
   getSchemaPartDescription(
     schemaPart?:
-      | import('json-schema').JSONSchema4
-      | import('json-schema').JSONSchema6
-      | import('json-schema').JSONSchema7
+      | (import('json-schema').JSONSchema4 & import('./validate').Extend)
+      | (import('json-schema').JSONSchema6 & import('./validate').Extend)
+      | (import('json-schema').JSONSchema7 & import('./validate').Extend)
       | undefined
   ): string;
   /**