+++ declarations/keywords/absolutePath.d.ts
@@ -1,5 +1,6 @@
 export default addAbsolutePathKeyword;
 export type Ajv = import('ajv').Ajv;
+export type ValidateFunction = import('ajv').ValidateFunction;
 export type SchemaUtilErrorObject = import('ajv').ErrorObject & {
   children?: import('ajv').ErrorObject[] | undefined;
 };