+++ declarations/util/Range.d.ts
@@ -40,4 +40,59 @@ declare namespace Range {
   };
 }
 type RangeValue = [number, boolean];
+/**
+ * @param {"left" | "right"} side
+ * @param {boolean} exclusive
+ * @returns {">" | ">=" | "<" | "<="}
+ */
+declare function getOperator(
+  side: 'left' | 'right',
+  exclusive: boolean
+): '>' | '>=' | '<' | '<=';
+/**
+ * @param {number} value
+ * @param {boolean} logic is not logic applied
+ * @param {boolean} exclusive is range exclusive
+ * @returns {string}
+ */
+declare function formatRight(
+  value: number,
+  logic: boolean,
+  exclusive: boolean
+): string;
+/**
+ * @param {number} value
+ * @param {boolean} logic is not logic applied
+ * @param {boolean} exclusive is range exclusive
+ * @returns {string}
+ */
+declare function formatLeft(
+  value: number,
+  logic: boolean,
+  exclusive: boolean
+): string;
+/**
+ * @param {number} start left side value
+ * @param {number} end right side value
+ * @param {boolean} startExclusive is range exclusive from left side
+ * @param {boolean} endExclusive is range exclusive from right side
+ * @param {boolean} logic is not logic applied
+ * @returns {string}
+ */
+declare function formatRange(
+  start: number,
+  end: number,
+  startExclusive: boolean,
+  endExclusive: boolean,
+  logic: boolean
+): string;
+/**
+ * @param {Array<RangeValue>} values
+ * @param {boolean} logic is not logic applied
+ * @return {RangeValue} computed value and it's exclusive flag
+ */
+declare function getRangeValue(
+  values: [number, boolean][],
+  logic: boolean
+): [number, boolean];
 type RangeValueCallback = (rangeValue: [number, boolean]) => boolean;