+++ declarations/validate.d.ts
@@ -3,10 +3,16 @@ export type JSONSchema4 = import('json-schema').JSONSchema4;
 export type JSONSchema6 = import('json-schema').JSONSchema6;
 export type JSONSchema7 = import('json-schema').JSONSchema7;
 export type ErrorObject = Ajv.ErrorObject;
+export type Extend = {
+  formatMinimum?: number | undefined;
+  formatMaximum?: number | undefined;
+  formatExclusiveMinimum?: boolean | undefined;
+  formatExclusiveMaximum?: boolean | undefined;
+};
 export type Schema =
-  | import('json-schema').JSONSchema4
-  | import('json-schema').JSONSchema6
-  | import('json-schema').JSONSchema7;
+  | (import('json-schema').JSONSchema4 & Extend)
+  | (import('json-schema').JSONSchema6 & Extend)
+  | (import('json-schema').JSONSchema7 & Extend);
 export type SchemaUtilErrorObject = Ajv.ErrorObject & {
   children?: Ajv.ErrorObject[] | undefined;
 };
@@ -29,13 +35,15 @@ export type ValidationErrorConfiguration = {
  */
 declare function validate(
   schema:
-    | import('json-schema').JSONSchema4
-    | import('json-schema').JSONSchema6
-    | import('json-schema').JSONSchema7,
-  options: any,
+    | (import('json-schema').JSONSchema4 & Extend)
+    | (import('json-schema').JSONSchema6 & Extend)
+    | (import('json-schema').JSONSchema7 & Extend),
+  options: object | object[],
   configuration?: ValidationErrorConfiguration | undefined
 ): void;
 declare namespace validate {
-  export { _default as ValidationError, _default as ValidateError };
+  export { ValidationError };
+  export { ValidationError as ValidateError };
 }
 import Ajv from 'ajv';
+import ValidationError from './ValidationError';